moving utils file inside utils folder and refactor
