+++ src/components/exchange/TxSummaryFixed.vue
@@ -479,7 +479,7 @@ import {
   TxStatusType,
   TxSummaryOrientation,
 } from '@/types';
-import { formatTxId, getBtcAddressExplorerUrl, getBtcTxExplorerUrl } from '@/services/utils';
+import { formatTxId, getBtcAddressExplorerUrl, getBtcTxExplorerUrl } from '@/utils/utils';
 import EnvironmentContextProviderService from '@/providers/EnvironmentContextProvider';
 import { State } from 'vuex-class';
 import { EnvironmentAccessorService } from '@/services/enviroment-accessor.service';
