+++ src/components/ledger/ConfirmLedgerTransaction.vue
@@ -191,7 +191,7 @@ import AdvancedData from '@/components/exchange/AdvancedData.vue';
 import SatoshiBig from '@/types/SatoshiBig';
 import LedgerTxBuilder from '@/middleware/TxBuilder/LedgerTxBuilder';
 import { WalletService } from '@/services';
-import { Machine } from '@/services/utils';
+import { Machine } from '@/utils/utils';
 import EnvironmentContextProviderService from '@/providers/EnvironmentContextProvider';
 import { PegInTxState } from '@/types/pegInTx';
 import * as constants from '@/store/constants';
