+++ src/middleware/TxBuilder/TrezorTxBuilder.ts
@@ -4,7 +4,7 @@ import {
   InputScriptType,
   NormalizedInput, NormalizedOutput, NormalizedTx, TrezorTx,
 } from '@/types';
-import { getAccountType } from '@/services/utils';
+import { getAccountType } from '@/utils/utils';
 import * as constants from '@/store/constants';
 import store from '../../store';
 import TxBuilder from './TxBuilder';
