+++ src/views/Home.vue
@@ -117,7 +117,7 @@ import { Action, State } from 'vuex-class';
 import SelectBitcoinWallet from '@/components/exchange/SelectBitcoinWallet.vue';
 import * as constants from '@/store/constants';
 import EnvironmentContextProviderService from '@/providers/EnvironmentContextProvider';
-import { isAllowedCurrentBrowser } from '@/services/utils';
+import { isAllowedCurrentBrowser } from '@/utils/utils';
 import LiqualityService from '@/services/LiqualityService';
 import {
   SessionState, TransactionType, PegInTxState, WalletAddress,
